Fix handling empty and too large dates in a save dialogue
There were two bugs in handling invalid dates in save dialogue.

(1) Years larger than 2038 caused this exception:

Date::Calc::PP::Date_to_Time(): Date::Calc::Date_to_Time(): date out of range at /usr/share/perl5/vendor_perl/Gscan2pdf/Document.pm line 2364

This a bug/feature in Date::Calc
<https://rt.cpan.org/Public/Bug/Display.html?id=102495>. Data::Calc
author imposed an artificial limit for case people pass a value which
could be out of range on some (32-bit) operating systems. This patch
handles this exception gracefully. I guess Date::Calc will be fixed
(i.e. lift the limit according to size of struct time_t) sooner or later
as year 2038 is getting closer. Be ware that there always will be dates
out of range even on fully 64-bit systems.

(2) An empty string was dereferenced as an array when handling the
exception. This patch implements a fall back for incomplete date string
in Gscan2pdf::Document::text_to_datetime() and let the metadata updater
to correct emptry string to today with that function. That guarantees
that exception will never get an empty string.

A side effect is that inputted large years (e.g 123456789-01-01) are
visually clamped to first 10 characters (123456789-). That can be
corrected in an independent patch which would remove the limit for the
date-time widget.

In independent question is what should do people who do not want to
store any date into the saved document. This use case has not been
supported. If the need emerges, the code should be changed to handle an
empty string as a valid, unset, value.

<https://sourceforge.net/p/gscan2pdf/bugs/407/>